pyproject.toml
[tool.poetry]
name = "chat-bot-api"
package-mode = false
version = "0.1.0"
description = ""
authors = ["Ahmed-Mohiuddin-Shah <mohiuddinlive@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
python-decouple = "^3.8"
uvicorn = "^0.30.1"
asyncio = "^3.4.3"
aiofiles = "^24.1.0"
bcrypt = "^4.1.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
pyjwt = "^2.8.0"
cachetools = "^5.4.0"
pydantic = "^2.9.2"
motor = "^3.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
